refactor(example): combine modules: command and example
1. combine modules: command and example

Closes https://github.com/openmessaging/dledger/issues/289
