--- command/src/main/java/io/openmessaging/storage/dledger/command/GetCommand.java
@@ -1,55 +1,0 @@
-/*
- * Copyright 2017-2022 The DLedger Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.openmessaging.storage.dledger.command;
-
-import com.alibaba.fastjson.JSON;
-import com.beust.jcommander.Parameter;
-import com.beust.jcommander.Parameters;
-import io.openmessaging.storage.dledger.client.DLedgerClient;
-import io.openmessaging.storage.dledger.entry.DLedgerEntry;
-import io.openmessaging.storage.dledger.protocol.GetEntriesResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Parameters(commandDescription = "Get data from DLedger server")
-public class GetCommand extends BaseCommand {
-
-    private static Logger logger = LoggerFactory.getLogger(GetCommand.class);
-
-    @Parameter(names = {"--group", "-g"}, description = "Group of this server")
-    private String group = "default";
-
-    @Parameter(names = {"--peers", "-p"}, description = "Peer info of this server")
-    private String peers = "n0-localhost:20911";
-
-    @Parameter(names = {"--index", "-i"}, description = "get entry from index")
-    private long index = 0;
-
-    @Override
-    public void doCommand() {
-        DLedgerClient dLedgerClient = new DLedgerClient(group, peers);
-        dLedgerClient.startup();
-        GetEntriesResponse response = dLedgerClient.get(index);
-        logger.info("Get Result:{}", JSON.toJSONString(response));
-        if (response.getEntries() != null && response.getEntries().size() > 0) {
-            for (DLedgerEntry entry : response.getEntries()) {
-                logger.info("Get Result index:{} {}", entry.getIndex(), new String(entry.getBody()));
-            }
-        }
-        dLedgerClient.shutdown();
-    }
-}
