+++ dledger/pom.xml
@@ -24,7 +24,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>dledger-core</artifactId>
+    <artifactId>dledger</artifactId>
 
     <dependencies>
         <dependency>
