+++ example/src/main/java/io/openmessaging/storage/dledger/example/appender/AppenderDLedger.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      https://www.apache.org/licenses/LICENSE-2.0
+ *        https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,44 +14,47 @@
  * limitations under the License.
  */
 
-package io.openmessaging.storage.dledger.example.register;
+package io.openmessaging.storage.dledger.example.appender;
 
+import com.alibaba.fastjson.JSON;
 import io.openmessaging.storage.dledger.DLedgerConfig;
-import io.openmessaging.storage.dledger.DLedgerServer;
-import io.openmessaging.storage.dledger.example.register.protocol.RegisterReadProcessor;
-import io.openmessaging.storage.dledger.example.register.protocol.RegisterWriteProcessor;
-import java.util.Arrays;
+import io.openmessaging.storage.dledger.proxy.DLedgerProxy;
+import java.util.Collections;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class RegisterDLedger {
+public class AppenderDLedger {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(RegisterDLedger.class);
+    private static Logger logger = LoggerFactory.getLogger(AppenderDLedger.class);
 
-    public static void bootstrap(DLedgerConfig dLedgerConfig) {
-        DLedgerServer dLedgerServer = new DLedgerServer(dLedgerConfig);
-        dLedgerServer.registerStateMachine(new RegisterStateMachine());
-        dLedgerServer.registerUserDefineProcessors(
-            Arrays.asList(new RegisterWriteProcessor(dLedgerServer), new RegisterReadProcessor(dLedgerServer))
-        );
-        dLedgerServer.startup();
-        LOGGER.info("RegisterDLedger started");
+    public static void bootstrapDLedger(List<DLedgerConfig> dLedgerConfigs) {
+        if (dLedgerConfigs == null || dLedgerConfigs.isEmpty()) {
+            logger.error("Bootstrap DLedger server error", new IllegalArgumentException("DLedgerConfigs is null or empty"));
+        }
+        DLedgerProxy dLedgerProxy = new DLedgerProxy(dLedgerConfigs);
+        dLedgerProxy.startup();
+        logger.info("DLedgers start ok with config {}", JSON.toJSONString(dLedgerConfigs));
         Runtime.getRuntime().addShutdownHook(new Thread(new Runnable() {
             private volatile boolean hasShutdown = false;
 
             @Override
             public void run() {
                 synchronized (this) {
-                    LOGGER.info("Shutdown hook was invoked");
+                    logger.info("Shutdown hook was invoked");
                     if (!this.hasShutdown) {
                         this.hasShutdown = true;
                         long beginTime = System.currentTimeMillis();
-                        dLedgerServer.shutdown();
+                        dLedgerProxy.shutdown();
                         long consumingTimeTotal = System.currentTimeMillis() - beginTime;
-                        LOGGER.info("Shutdown hook over, consuming total time(ms): {}", consumingTimeTotal);
+                        logger.info("Shutdown hook over, consuming total time(ms): {}", consumingTimeTotal);
                     }
                 }
             }
         }, "ShutdownHook"));
+    }
+
+    public static void bootstrapDLedger(DLedgerConfig dLedgerConfig) {
+        bootstrapDLedger(Collections.singletonList(dLedgerConfig));
     }
 }
