+++ example/src/main/java/io/openmessaging/storage/dledger/example/appender/command/AppendCommand.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2017-2022 The DLedger Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.openmessaging.storage.dledger.example.appender.command;
+
+import com.alibaba.fastjson.JSON;
+import com.beust.jcommander.Parameter;
+import com.beust.jcommander.Parameters;
+import io.openmessaging.storage.dledger.client.DLedgerClient;
+import io.openmessaging.storage.dledger.example.common.command.BaseCommand;
+import io.openmessaging.storage.dledger.protocol.AppendEntryResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+@Parameters(commandDescription = "Append data to AppenderDLedger")
+public class AppendCommand extends BaseCommand {
+
+    private static Logger logger = LoggerFactory.getLogger(AppendCommand.class);
+
+    @Parameter(names = {"--group", "-g"}, description = "Group of this server")
+    private String group = "default";
+
+    @Parameter(names = {"--peers", "-p"}, description = "Peer info of this server")
+    private String peers = "n0-localhost:20911";
+
+    @Parameter(names = {"--data", "-d"}, description = "the data to append")
+    private String data = "Hello";
+
+    @Parameter(names = {"--count", "-c"}, description = "append several times")
+    private int count = 1;
+
+    @Override
+    public void doCommand() {
+        DLedgerClient dLedgerClient = new DLedgerClient(group, peers);
+        dLedgerClient.startup();
+        for (int i = 0; i < count; i++) {
+            AppendEntryResponse response = dLedgerClient.append(data.getBytes());
+            logger.info("Append Result:{}", JSON.toJSONString(response));
+        }
+        dLedgerClient.shutdown();
+    }
+}
+
