+++ example/src/main/resources/appender/config.example.error.yaml
