--- jepsen/nodes
@@ -1,5 +1,0 @@
-n0
-n1
-n2
-n3
-n4
