+++ pom.xml
@@ -52,7 +52,6 @@
 
 
     <modules>
-        <module>command</module>
         <module>dledger</module>
         <module>proxy</module>
         <module>example</module>
@@ -67,12 +66,7 @@
             </dependency>
             <dependency>
                 <groupId>io.openmessaging.storage</groupId>
-                <artifactId>dledger-core</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>io.openmessaging.storage</groupId>
-                <artifactId>dledger-command</artifactId>
+                <artifactId>dledger</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
