+++ proxy/pom.xml
@@ -15,7 +15,7 @@
     <dependencies>
         <dependency>
             <groupId>io.openmessaging.storage</groupId>
-            <artifactId>dledger-core</artifactId>
+            <artifactId>dledger</artifactId>
         </dependency>
         <dependency>
             <groupId>org.yaml</groupId>
